+++ src/main/java/com/dxctraining/flight/dao/IFlightDao.java
@@ -1,16 +1,14 @@
 package com.dxctraining.flight.dao;
 
+import java.math.BigInteger;
 import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.dxctraining.flight.entities.Flight;
 
 public interface IFlightDao {
 	Flight addFlight(Flight flight);
 	Flight modifyFlight(Flight flight);
-	Flight viewByFlightNum(Integer flightNum);
+	Flight viewByFlightNum(BigInteger flightNum);
 	List<Flight> viewAllFlights();
-	void delete(Integer flightNum);
-}
+	void delete(BigInteger flightNum);
+	}
