+++ src/main/java/com/dxctraining/flight/service/IFlightService.java
@@ -1,5 +1,6 @@
 package com.dxctraining.flight.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.dxctraining.flight.entities.Flight;
@@ -7,9 +8,9 @@ import com.dxctraining.flight.entities.Flight;
 public interface IFlightService {
 	Flight addFlight(Flight flight);
 	Flight modifyFlight(Flight flight);
-	Flight viewByFlightNum(Integer flightNum);
+	Flight viewByFlightNum(BigInteger flightNum);
 	List<Flight> viewAllFlights();
-	void delete(Integer flightNum);
+	void delete(BigInteger flightNum);
 	void validate(Object obj);
 
 }
